+++ Tests/IceCreamTests/SyncEngineTests.swift
@@ -81,77 +81,3 @@ class SyncEngineTests: XCTestCase {
         wait(for: [expectation], timeout: 0.1)
     }
 }
-
-private func mockCKRecord() -> CKRecord {
-    return CKRecord(recordType: "Test", recordID: mockCKRecordId())
-}
-
-private func mockCKRecordId() -> CKRecordID {
-    return CKRecordID(recordName: "RecordId", zoneID: CKRecordZoneID(zoneName: "Test", ownerName: "Test"))
-}
-
-class MockObject: Object {
-    var registerLocalDatabaseCalled: (() -> Void)? = nil
-    var addCalled: ((CKRecord) -> Void)? = nil
-    var removeCalled: ((CKRecordID) -> Void)? = nil
-
-    var recordType: String = "Test"
-
-    var customZoneID: CKRecordZoneID = CKRecordZoneID(zoneName: "Test", ownerName: "Test")
-
-    var zoneChangesToken: CKServerChangeToken? = nil
-
-    var isCustomZoneCreated: Bool = true
-
-    func registerLocalDatabase() {
-        registerLocalDatabaseCalled?()
-    }
-
-    func cleanUp() { }
-
-    func add(record: CKRecord) {
-        addCalled?(record)
-    }
-
-    func delete(recordID: CKRecordID) {
-        removeCalled?(recordID)
-    }
-
-    var pipeToEngine: (([CKRecord], [CKRecordID]) -> ())?
-}
-
-extension MockObject: Syncable { }
-
-extension MockObject: CKRecordConvertible, CKRecordRecoverable {
-    var isDeleted: Bool {
-        return false
-    }
-}
-
-class CloudKitMock: CloudKitDataSourcing {
-    var mockAddedRecords: [CKRecord] = []
-    var mockRemovedRecordIds: [CKRecordID] = []
-    var recordsSentToStore: [CKRecord] = []
-    var recordIdsSendToDelete: [CKRecordID] = []
-    var fetchChangesCalled: (() -> Void)? = nil
-
-    func cloudKitAvailable(_ completed: @escaping (Bool) -> Void) {
-        completed(true)
-    }
-
-    func fetchChanges(recordZoneTokenUpdated: @escaping (CKRecordZoneID, CKServerChangeToken?) -> Void, added: @escaping ((CKRecord) -> Void), removed: @escaping ((CKRecordID) -> Void)) {
-        fetchChangesCalled?()
-        for record in mockAddedRecords {
-            added(record)
-        }
-
-        for id in mockRemovedRecordIds {
-            removed(id)
-        }
-    }
-
-    func syncRecordsToCloudKit(recordsToStore: [CKRecord], recordIDsToDelete: [CKRecordID], completion: ((Error?) -> ())?) {
-        recordsSentToStore = recordsToStore
-        recordIdsSendToDelete = recordIDsToDelete
-    }
-}
